Add support for closed captions.
If a file matching the video filename but ending in ".vtt" (or with ".vtt" tacked onto the full video filename with extension) is in the same directory,
Channel Two will show closd captions for that video if the captions=true flag has been set in the schedule file or if you press 'c' while Channel Two is
open.

If you press 'c' to force captions on, they'll be shown for every video until you press 'c' again, at which point only videos that have been designated in
the schedule as having captions enabled will show captions.

Fixes #2.